Do not use mobile detect bundle anymore
The mobile detect bundle provides a cookie based method to switch the
view. It's not required, because Contao provides it's own method. And
for cache optimization the extra cookie should not be set.

As it's not possible to disable the cookie in the bundle, we use the
mobile detect library directly.